fitst version of Project 1 exercise b)
first version of the program to solve the given set of linear equations
and also to plot the (used gnuplot and gnuplot- skript). We have to look
at the code and try to make the program faster ... I haven't done all
coments on the program yet, but I will do this later. As far as I did
it, I can tell, that the first solution fits all in all to the explicit
solution. Just have a look at the png- file.